fw: Break out htif functions for qemu to separate files
