+++ hw/application_fpga/Makefile
@@ -115,7 +115,8 @@ FIRMWARE_DEPS = \
 	$(P)/fw/tk1/flash.h \
 	$(P)/fw/tk1/partition_table.h \
 	$(P)/fw/tk1/preload_app.h \
-	$(P)/fw/tk1/auth_app.h
+	$(P)/fw/tk1/auth_app.h \
+	$(P)/fw/tk1/htif.h
 
 FIRMWARE_OBJS = \
 	$(P)/fw/tk1/main.o \
@@ -129,7 +130,8 @@ FIRMWARE_OBJS = \
 	$(P)/fw/tk1/flash.o \
 	$(P)/fw/tk1/partition_table.o \
 	$(P)/fw/tk1/preload_app.o \
-	$(P)/fw/tk1/auth_app.o
+	$(P)/fw/tk1/auth_app.o \
+	$(P)/fw/tk1/htif.o
 
 FIRMWARE_SOURCES = \
 	$(P)/fw/tk1/main.c \
@@ -142,7 +144,8 @@ FIRMWARE_SOURCES = \
 	$(P)/fw/tk1/flash.c \
 	$(P)/fw/tk1/partition_table.c \
 	$(P)/fw/tk1/preload_app.c \
-	$(P)/fw/tk1/auth_app.c
+	$(P)/fw/tk1/auth_app.c \
+	$(P)/fw/tk1/htif.c
 
 TESTFW_OBJS = \
 	$(P)/fw/testfw/main.o \
